requirements.txt
discord.py
flask
feedparser
aiohttp
beautifulsoup4